trolloBackend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.7.0
boto3==1.29.0
botocore==1.32.0
Django==4.2.7
django-storages==1.14.2
djangorestframework==3.14.0
dynamorm==0.11.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
jmespath==1.0.1
marshmallow==3.20.1
packaging==23.2
promise==2.3
pynamodb==5.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
Rx==1.6.3
s3transfer==0.7.0
setuptools==68.2.2
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.0.7
wrapt==1.16.0
